Refactored tests. Split them into multiple files for increased readability
